requirements.txt
asana==0.10.3
certifi==2021.10.8
charset-normalizer==2.0.7
idna==3.2
oauthlib==3.1.1
python-dotenv==0.19.1
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
urllib3==1.26.7